New Construction: Automatic Portable Turrets
Items required to construct: any energy gun, a proximity sensor, a screwdriver, wrench, a welder and a total of 9 metal.

This turret can best be described as a mixture of an ED-209 and an AI turret. It fires the same projectile as the gun it uses, and has a few customizable behaviours for people with security access to play with. The turrets can be unbolted and dragged off to be deployed anywhere, but can not turn on and shoot unless it's bolted into place. Can be controlled by the AI and cyborgs, but isn't controlled by /area's in the same way normal AI turrets are.

This will probably be expanded on in the future if I or one of the other coders want it to be expanded.




git-svn-id: http://tgstation13.googlecode.com/svn/trunk@1746 316c924e-a436-60f5-8080-3fe189b3f50e