Delicious Nuka cola is back, with a bottled sprite for its glass (don't question why it turns a glass into a bottle and back again, it is just physics).
Seriously don't remove ice cups from the booze O mat, it is meant to use a cheap Styrofoam cup as its sprite (that's the look I was going for when I made it, after all.. what else would you use?).

Synaptizine is now useful. It reduces weakness/ stun/ silence/ drowsiness/ paralysis to zero. Meaning you can't be stunned, or knocked down for more than a moment. it is great for when a rogue borg rolls around or if you need to revive some one who was stunned by a shocked door (but not killed by it) or if you just need to get away from those dang security officers. It'll do more too, but it is pretty much anti-zombie powder (just it can't cancel the toxic effects of zombie powder).

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@2143 316c924e-a436-60f5-8080-3fe189b3f50e