added a proc for feedback feedback_add_details(var/variable,var/details)
This adds the contents of the details variable at the end of the variable, which is defined in the variable variable. You know this last line makes perfect sense.

It also adds a space between what is already in details and what you're adding.

in effect it's a 

s += " content"

Added an autogibber. If you bump into it from the right end, it gibs you.

Added an exit into space at engineering, by the escape pod.
- External airlocks require engineering + external access
- pod airlocks don't require any access

Added an exit into space at the pod in security
- External airlocks require security + external access


git-svn-id: http://tgstation13.googlecode.com/svn/trunk@2698 316c924e-a436-60f5-8080-3fe189b3f50e