Examining humans is now smarter and displays only the things you can see. Certain outerwear suits and helmets hide other pieces of clothing on you.
A diagram that displays what can hide what:
http://www.kamletos.si/inventory%20visibility%20hierarchy.png

Note that not all suits and helmets hide stuff.

Added this to the changelog.

Technical information:

Renamed gimmick.dm to z_gimmick.dm because whenever I doubleclicked on any piece of clothing in the object tree it always pointed me to some random gimmick item instead of the root definition. Now the file is at the bottom of the folder and so everything else is looked up first.

Created a bitflag variable which is used to determine which piece of clothing hides another. The variable is flags_inv and the flags it contains are:

#define HIDEGLOVES 1		//APPLIES ONLY TO THE EXTERIOR SUIT!!
#define HIDESUITSTORAGE 2	//APPLIES ONLY TO THE EXTERIOR SUIT!!
#define HIDEJUMPSUIT 4		//APPLIES ONLY TO THE EXTERIOR SUIT!!
#define HIDESHOES 8			//APPLIES ONLY TO THE EXTERIOR SUIT!!
#define HIDEMASK 1			//APPLIES ONLY TO HELMETS!!
#define HIDEEARS 2			//APPLIES ONLY TO HELMETS!!
#define HIDEEYES 4			//APPLIES ONLY TO HELMETS!!

The first 4 only apply to exterior suits and the last 3 only to helmets, so they can use the same numbers. 

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@3031 316c924e-a436-60f5-8080-3fe189b3f50e