Added a partially constructed teleporter to the syndicate (nuclear) shuttle. It is behind a blast door operated by the same computer that opens the outer airlock to prevent the operatives from using it will still on the centcomm Z level.
A teleporter circuit board has been added to the nuke radio, and can be purchased for 20 telecrystals. This will allow them to finish the teleporter onboard their ship. This will be probably be absurdly overpowered and need to be reverted, but no harm in trying it out, right?

Moved the teleporter board out secure tech and into the RD's office. Because teleporters cannot be built from scratch, the board serves almost no purpose but to be deconstructed anyway. And having it in the RD's office will prevent nuke teams from stealing it from space to complete their machine.

Replaced the mechanical toolboxes on the nuke shuttle with syndicate toolboxes.

Updated the changelog

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@3076 316c924e-a436-60f5-8080-3fe189b3f50e