Shades, Juggernauts, Wraiths, Artificers all now have minds.
If a cultist builds a construct, that construct will be a cultist. This means they can see cultists finally (and will probably count for the escape objective, so that might need changing numbers wise)

Fixed some typos in the cult runes (godamnit Urist) and the blind rune can no longer be made into a talisman, cause a handheld, instant, AoE blind that lasts 20 seconds with no counter is retarded.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@3507 316c924e-a436-60f5-8080-3fe189b3f50e