Committing my server autojoiner / traffic regulator code. It's a small BYOND application you can host that will automatically redirect players to the most desireable server (currently, it will only compare player quantities and select the least-populated server). It's not finished yet, but it works for the most part.
This is meant to be hosted on the same machine as the servers are hosted on, otherwise the list of admins will not be able to be generated.

Be sure to look at config.txt to under the Redirector directory. It's pre-configured but can be modified for other servers' usage.



git-svn-id: http://tgstation13.googlecode.com/svn/trunk@3961 316c924e-a436-60f5-8080-3fe189b3f50e