- Added UI buttons which appear only when you have certain items on you to make toggling some stuff easier and lessen the need to juggle a million items all the time.
Screenshot: (Upper left)
http://www.kamletos.si/object%20ui%20action%20buttons.png

There can be a maximum of 5 of these buttons shown at any one time.

Currently the following items have a UI button associated with them:
- Flashlights
- Hardhat
- RIG suits
- Welding masks
- Jetpacks

To add more buttons, add a button icon_state to icons/mob/screen1_action.dmi, then set the variable icon_action_button to the icon state of the icon you just added. This is all you need to do if you want the default action to be taken when the button is clicked, which is attack_self(). If you don't want the default, you can redefine the ui_action_click() proc. A call of ..() will call attack_self(), so mind when you use that.

git-svn-id: http://tgstation13.googlecode.com/svn/trunk@4193 316c924e-a436-60f5-8080-3fe189b3f50e