Welding gas mask
Adds a new item to RnD which is a gasmask with integrated welding visor.
Requires low level (2) of Engineering and Materials, 4000 metal, 2000 glass (1000 more of each compared to a standard welding helmet), fits in the mask slot.

The idea behind this is that it will give engineers the functionality of Rig Helmets (head lamp and welding visor, and free eye slot for mesons) when combined with a hard hat.

Included some placeholder sprites:
https://dl.dropboxusercontent.com/u/95696802/Commits/weldinggasmask.png