Fixes temperature guns for the doubly last time
Makes the temperature projectile get its temperature from the gun, using
a proc used to give the bullet properties based on the properties of the
gun, which can be used with other projectiles for different guns.
Also names the temperature projectile based on its temperature.  500
degrees is not a freeze beam, that's just silly.