Removed all the spawn(0) from bump code and the special check for turfs that rolan7 made a while ago on the mob different Bump() procs. Added a special check on move() code, only when you're pulling something. If the loc of what you're pulling is the same after you moved, you'll execute the pulling. Otherwise you'll stop pulling. This is to avoid pushing something that you're pulling and moving it back, looking like you didn't do anything. This was "handled" by the spawn(0) procs.
Reworked cyborg recharge stations, they aren't part of the death() and Bump() procs of cyborgs and they act like lockers. Click them to open them and again to close them. Added a sprite for when it's closed with no occupant inside (it's just the closed sprite with red lights instead of a blinking green light).